requirements.txt
anyio==4.0.0
async-tkinter-loop==0.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Pillow==9.5.0
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.4
